requirements.txt
more_itertools
discord.py
requests
json